Refactor: 추가 사항 반영
